/MT build and import from unicode path
 - .pyc file written to correct place.
 - Loading packages working
 - MT build issues fixed for BZ2 module

BZ2 module opens the file with the Python File() constructor, but writes to the
FILE* directly with fwrite().  When built with /MT, the fopen() and fwrite() are
then called from separate copies of the C Runtime, hence the FILE* is no longer
valid.  This commit introduces the fix, which opens the file directly in the BZ2
module, and then creates a PyFileObject from the FILE*.

Minor fix to pydoc, so unicode paths won't crash the help() call.

Simply skips displaying the module contents if it errors.
This really needs a better fix.